refactor: 에러와 예외 도메인별 분리
