+++ src/main/java/org/sopt/article/service/ArticleService.java
@@ -5,12 +5,13 @@ import org.sopt.article.dto.request.ArticleCreateRequest;
 import org.sopt.article.dto.response.ArticleListResponse;
 import org.sopt.article.dto.response.ArticleResponse;
 import org.sopt.article.entity.Article;
+import org.sopt.article.exception.ArticleErrorCode;
 import org.sopt.article.repository.ArticleRepository;
 import org.sopt.global.exception.BusinessException;
-import org.sopt.global.exception.ErrorCode;
-import org.sopt.global.exception.domain.article.ArticleException;
-import org.sopt.global.exception.domain.member.MemberException;
+import org.sopt.article.exception.ArticleException;
+import org.sopt.member.exception.MemberException;
 import org.sopt.member.entity.Member;
+import org.sopt.member.exception.MemberErrorCode;
 import org.sopt.member.repository.MemberRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -31,7 +32,7 @@ public class ArticleService  {
         validateTitleExists(request.title());
 
         Member member = memberRepository.findById(request.memberId())
-                .orElseThrow(() -> new MemberException(ErrorCode.MEMBER_NOT_FOUND));
+                .orElseThrow(() -> new MemberException(MemberErrorCode.MEMBER_NOT_FOUND));
 
         Article article = Article.create(request.title(), request.content(),LocalDate.now(),request.tag(),member);
 
@@ -44,7 +45,7 @@ public class ArticleService  {
     public ArticleResponse findArticle(Long articleId) {
 
         Article article = articleRepository.findById(articleId)
-                .orElseThrow(() -> new ArticleException(ErrorCode.ARTICLE_NOT_FOUND));
+                .orElseThrow(() -> new ArticleException(ArticleErrorCode.ARTICLE_NOT_FOUND));
 
         return ArticleResponse.from(article);
 
@@ -68,7 +69,7 @@ public class ArticleService  {
     private void validateTitleExists(String title) {
 
         if(articleRepository.existsByTitle(title)){
-            throw new BusinessException(ErrorCode.DUPLICATE_TITLE);
+            throw new ArticleException(ArticleErrorCode.DUPLICATE_TITLE);
         }
 
     }
