+++ src/main/java/org/sopt/global/exception/BusinessException.java
@@ -1,9 +1,9 @@
 package org.sopt.global.exception;
 
-public class BusinessException extends RuntimeException {
+public abstract class BusinessException extends RuntimeException {
     private final ErrorCode errorCode;
 
-    public BusinessException(ErrorCode errorCode) {
+    protected BusinessException(ErrorCode errorCode) {
         super(errorCode.getMessage());
         this.errorCode = errorCode;
     }
