+++ src/main/java/org/sopt/member/entity/Member.java
@@ -6,8 +6,8 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.sopt.article.entity.Article;
-import org.sopt.global.exception.ErrorCode;
-import org.sopt.global.exception.domain.member.MemberException;
+import org.sopt.member.exception.MemberException;
+import org.sopt.member.exception.MemberErrorCode;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
@@ -47,7 +47,7 @@ public class Member{
     private static void validateAge(String birth){
         int age = LocalDate.now().getYear() - LocalDate.parse(birth).getYear();
         if( age < 20){
-            throw new MemberException(ErrorCode.AGE_LOW);
+            throw new MemberException(MemberErrorCode.AGE_LOW);
         }
     }
 }
