+++ src/main/java/org/sopt/member/exception/MemberException.java
@@ -1,4 +1,4 @@
-package org.sopt.global.exception.domain.member;
+package org.sopt.member.exception;
 
 import org.sopt.global.exception.BusinessException;
 import org.sopt.global.exception.ErrorCode;
